3496 - Rename Sections folder to GeoSection
